Adding retrying module dependency to KIA deployment script
Summary:
# Context
In D50562498 we added retrying logic to key generation in KIA. During development and local testing we build the project using buck, however for AWS deployment we use terraform (with docker image).
For the new dependency added here (retrying module). We also need to add the same to the Dockerfile so that it can be included in the KIA deployment.

# Changes in the diff
Added retrying module install step in Dockerfile for KIA.

Differential Revision:
D51490362

Privacy Context Container: L416713